requirements.txt
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask==3.1.0
gunicorn==23.0.0
marshmallow==3.25.1
numpy==2.2.1
pandas==2.2.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2024.2
simplejson==3.19.3
xarray==2024.11.0
